+++ src/util.ts
@@ -1,6 +1,8 @@
+import chalk from "chalk";
 import { spawn } from "child_process";
 import fs from "fs-extra";
 import path from "path";
+import validateProjectName from "validate-npm-package-name";
 
 /**
  * Executes a command in a spawned process.
@@ -73,4 +75,30 @@ export const valueReplacer = (
     await fs.writeFile(filePath, newFileContent, "utf8");
     return;
   });
+};
+
+/**
+ * Validates the application name according to NPM naming conventions.
+ * @param applicationName Name of application.
+ */
+export const validateApplicationName = (applicationName: any) => {
+  const validation = validateProjectName(applicationName);
+  if (!validation.validForNewPackages) {
+    console.error(
+      `Cannot create an application named ${chalk.red(
+        `"${applicationName}"`
+      )} because of npm naming restrictions:`
+    );
+    console.log("");
+
+    [...(validation.errors || []), ...(validation.warnings || [])].forEach(
+      (error) => {
+        console.error(chalk.red(`  * ${error}`));
+      }
+    );
+
+    console.log("");
+    console.error("Please choose a different application name.");
+    process.exit(1);
+  }
 };