+++ CHANGELOG.md
@@ -9,10 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Interactive option. Use the flag `--interactive` to use this mode.
+- Interactive option. Use the flag `--interactive` to use this mode
 - Comments :)
 - Cleanup on global error
 - Option for Author Name
+- Validation of application name according to NPM conventions
 
 ### Changed